feat: 예외상황 문서화
