+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -3,12 +3,13 @@ package roomescape.exception;
 import static roomescape.exception.ExceptionType.INVALID_DATE_TIME_FORMAT;
 import static roomescape.exception.ExceptionType.UNEXPECTED_ERROR;
 
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import java.io.PrintWriter;
 import java.io.StringWriter;
-import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -18,25 +19,12 @@ import roomescape.dto.ErrorResponse;
 @ControllerAdvice
 public class GlobalExceptionHandler {
 
-    private static final Set<String> SERVER_ERROR_CODES = Set.of(
-            "INVALID_API_KEY",
-            "NOT_FOUND_TERMINAL_ID",
-            "INVALID_AUTHORIZE_AUTH",
-            "INVALID_UNREGISTERED_SUBMALL",
-            "UNAPPROVED_ORDER_ID",
-            "UNAUTHORIZED_KEY",
-            "FORBIDDEN_REQUEST",
-            "INCORRECT_BASIC_AUTH_FORMAT",
-            "FAILED_PAYMENT_INTERNAL_SYSTEM_PROCESSING",
-            "FAILED_INTERNAL_SYSTEM_PROCESSING",
-            "UNKNOWN_PAYMENT_ERROR"
-    );
-
     private final Logger logger = LoggerFactory.getLogger(getClass());
     private final StringWriter stringWriter = new StringWriter();
     private final PrintWriter printWriter = new PrintWriter(stringWriter);
 
     @ExceptionHandler(HttpMessageNotReadableException.class)
+    @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     public ResponseEntity<ErrorResponse> handle(HttpMessageNotReadableException e) {
         logger.error(getStackTrace(e));
         return ResponseEntity.status(INVALID_DATE_TIME_FORMAT.getStatus())
@@ -44,6 +32,10 @@ public class GlobalExceptionHandler {
     }
 
     @ExceptionHandler(RoomescapeException.class)
+    @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+    @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+    @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+    @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     public ResponseEntity<ErrorResponse> handle(RoomescapeException e) {
         logger.error(getStackTrace(e));
         return ResponseEntity
@@ -51,20 +43,8 @@ public class GlobalExceptionHandler {
                 .body(new ErrorResponse(e.getMessage()));
     }
 
-    @ExceptionHandler(PaymentException.class)
-    public ResponseEntity<ErrorResponse> handle(PaymentException e) {
-        logger.error(getStackTrace(e));
-        if (SERVER_ERROR_CODES.contains(e.getCode())) {
-            return ResponseEntity
-                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(new ErrorResponse(e.getMessage()));
-        }
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(new ErrorResponse(e.getMessage()));
-    }
-
     @ExceptionHandler(Exception.class)
+    @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     public ResponseEntity<ErrorResponse> handle(Exception e) {
         logger.error(getStackTrace(e));
         return ResponseEntity
